Add FakeCrashFileMarker
