+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/fakes/injection/FakeCrashModule.kt
@@ -2,14 +2,14 @@ package io.embrace.android.embracesdk.fakes.injection
 
 import io.embrace.android.embracesdk.fakes.FakeCrashService
 import io.embrace.android.embracesdk.injection.CrashModule
-import io.embrace.android.embracesdk.internal.crash.CrashFileMarker
+import io.embrace.android.embracesdk.internal.crash.CrashFileMarkerImpl
 import io.embrace.android.embracesdk.internal.crash.LastRunCrashVerifier
 import io.embrace.android.embracesdk.samples.AutomaticVerificationExceptionHandler
 import io.mockk.mockk
 
 internal class FakeCrashModule : CrashModule {
     override val lastRunCrashVerifier = LastRunCrashVerifier(
-        CrashFileMarker(mockk(relaxed = true), mockk(relaxed = true)),
+        CrashFileMarkerImpl(mockk(relaxed = true), mockk(relaxed = true)),
         mockk(relaxed = true)
     )
     override val crashService = FakeCrashService()
